Rename new unified MOTECH connection management command.
Performed as separate commit for diff clarity.
